requirements.txt
streamlit
pillow
anthropic